Course Timetabling: Database loader  - added ability to skip purging invalid placements phase (when solver parameter General.PurgeInvalidPlacements is set to false, defaults to true)

+++ trunk/JavaSource/org/unitime/timetable/solver/TimetableDatabaseLoader.java
@@ -2936,7 +2936,8 @@ public class TimetableDatabaseLoader extends TimetableLoader {
             }
         }
 
-		purgeInvalidValues();
+		if (getModel().getProperties().getPropertyBoolean("General.PurgeInvalidPlacements", true))
+			purgeInvalidValues();
 		
 		for (Constraint c: getModel().constraints()) {
 			if (c instanceof SpreadConstraint)
